Move print_syntax_diff to compare.py
