+++ diffkemp/compare.py
@@ -1,4 +1,3 @@
-from diffkemp.diffkemp import print_syntax_diff
 from diffkemp.config import Config
 from diffkemp.utils import CMP_OUTPUT_FILE
 from diffkemp.llvm_ir.llvm_module import LlvmParam, LlvmModule
@@ -107,6 +106,7 @@ def compare(args):
             new_fun_desc.mod.clean_module()
             LlvmModule.clean_all()
 
+    result.graph = result_graph
     # Create yaml output
     if output_dir is not None and os.path.isdir(output_dir):
         _create_yaml_output(args, result, output_dir)
@@ -241,3 +241,89 @@ def _print_stats(errors, result, extended_stat):
     print(f"\nStatistics\n{'-' * 11}")
     result.stop_time = default_timer()
     result.report_stat(errors, extended_stat)
+
+
+def print_syntax_diff(snapshot_dir_old, snapshot_dir_new, fun, fun_result,
+                      fun_tag, output_dir, show_diff, full_diff,
+                      initial_indent):
+    """
+    Log syntax diff of 2 functions. If log_files is set, the output is printed
+    into a separate file, otherwise it goes to stdout.
+    :param snapshot_dir_old: Old snapshot directory.
+    :param snapshot_dir_new: New snapshot directory.
+    :param fun: Name of the analysed function
+    :param fun_tag: Analysed function tag
+    :param fun_result: Result of the analysis
+    :param output_dir: True if the output is to be written into a file
+    :param show_diff: Print syntax diffs.
+    :param full_diff: Print semantics-preserving syntax diffs too.
+    :param initial_indent: Initial indentation of printed messages
+    """
+    def text_indent(text, width):
+        """Indent each line in the text by a number of spaces given by width"""
+        return ''.join(" " * width + line for line in text.splitlines(True))
+
+    old_dir_abs_path = os.path.join(os.path.abspath(snapshot_dir_old), "")
+    new_dir_abs_path = os.path.join(os.path.abspath(snapshot_dir_new), "")
+
+    if fun_result.kind == Result.Kind.NOT_EQUAL or (
+            full_diff and any([x.diff for x in fun_result.inner.values()])):
+        if output_dir:
+            output = open(os.path.join(output_dir, "{}.diff".format(fun)), "w")
+            output.write(
+                "Found differences in functions called by {}".format(fun))
+            if fun_tag is not None:
+                output.write(" ({})".format(fun_tag))
+            output.write("\n\n")
+            indent = initial_indent + 2
+        else:
+            output = sys.stdout
+            if fun_tag is not None:
+                output.write(text_indent("{} ({}):\n".format(fun, fun_tag),
+                                         initial_indent))
+            else:
+                output.write(text_indent("{}:\n".format(fun), initial_indent))
+            indent = initial_indent + 4
+
+        for called_res in sorted(fun_result.inner.values(),
+                                 key=lambda r: r.first.name):
+            if called_res.diff == "" and called_res.first.covered:
+                # Do not print empty diffs
+                continue
+
+            output.write(
+                text_indent("{} differs:\n".format(called_res.first.name),
+                            indent - 2))
+            if not output_dir:
+                output.write(text_indent("{{{\n", indent - 2))
+
+            if called_res.first.callstack:
+                output.write(
+                    text_indent("Callstack ({}):\n".format(snapshot_dir_old),
+                                indent))
+                output.write(text_indent(
+                    called_res.first.callstack.as_str_with_rel_paths(
+                        old_dir_abs_path), indent))
+                output.write("\n\n")
+            if called_res.second.callstack:
+                output.write(
+                    text_indent("Callstack ({}):\n".format(snapshot_dir_new),
+                                indent))
+                output.write(text_indent(
+                    called_res.second.callstack.as_str_with_rel_paths(
+                        new_dir_abs_path), indent))
+                output.write("\n\n")
+
+            if show_diff:
+                if (called_res.diff.strip() == "" and
+                        called_res.macro_diff is not None):
+                    output.write(text_indent(
+                        "\n".join(map(str, called_res.macro_diff)), indent))
+                else:
+                    output.write(text_indent("Diff:\n", indent))
+                    output.write(text_indent(
+                        called_res.diff, indent))
+
+            if not output_dir:
+                output.write(text_indent("}}}\n", indent - 2))
+            output.write("\n")
